Split js tests in 3 subgroups.
js test are the longer ~8 minutes vs ~3 for Python, and fails the more often.
Hopefully this will speedup the build and have less test relaunched when we
have a random faillure.  Splitting in 3 should make the js test a bit faster
and gave us some rooms without increasing the number of groups.