First step in reintegrating Jedi
If Jedi is installed expose a private API use it with prompt toolkit.
Jedi does not _yet_ provide all the completion IPython has, so this is
still a bit awkward. In order to debug this (and see what is Jedi
provided we for now inject a fake Jedi/IPython delimiter in the menu.

Next step would to tag completions we don't expect Jedi to have and not
mark them as "missing".

Jedi completion and this behavior are enabled by default, but could
likely be opt-in.